ref: deepen output planning boundary
