+++ src-tauri/src/commands/audio_processing.rs
@@ -1,10 +1,7 @@
 use crate::audio;
 use crate::audio::file_list::FileListInfo;
 use crate::audio::job_registry::{CancellationChecker, JobId};
-use crate::audio::output_path::{
-    build_output_path_preview, resolve_output_plan, CollisionPolicy, OutputKind,
-    PlannedOutputAction, ResolvedOutputPlan,
-};
+use crate::audio::output_path::{OutputKind, PlannedOutputAction, ResolvedOutputPlan};
 use crate::audio::settings_encoder::{
     resolve_encoder_type, validate_encoder_settings, validate_requested_encoder_available,
     EncoderType,
@@ -15,40 +12,23 @@ use crate::audio::toolchain::{
 };
 use crate::audio::{QueueEvent, QueueItem};
 use crate::commands::audio_types::{
-    JobType, OutputNamingConfig, ProcessCommandResult, ProcessPayload, ProcessResultEntry,
-    ProcessResultStatus, ProcessingPreflightPlan,
+    JobType, ProcessCommandResult, ProcessPayload, ProcessResultEntry, ProcessResultStatus,
+    ProcessingPreflightPlan,
 };
 use crate::errors::sanitize_path_for_display;
 use crate::errors::{AppError, AppErrorEnvelope, Result};
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 use std::future::Future;
 use std::path::{Path, PathBuf};
 use std::pin::Pin;
 use tauri::Emitter;
 use tokio::sync::OwnedSemaphorePermit;
 
-struct ProcessingInputs {
-    output_naming: OutputNamingConfig,
-    base_output_dir: PathBuf,
-    preview_seconds: Option<f64>,
-}
-
-#[derive(Debug, Clone)]
-struct PlannedProcessingJob {
-    input_index: Option<usize>,
-    input_path: Option<PathBuf>,
-    output: ResolvedOutputPlan,
-    metadata: Option<crate::metadata::AudiobookMetadata>,
-}
-
-#[derive(Debug, Clone)]
-struct ResolvedProcessingPlan {
-    job_type: JobType,
-    preview_seconds: Option<f64>,
-    collision_policy: CollisionPolicy,
-    plan_signature: String,
-    jobs: Vec<PlannedProcessingJob>,
-}
+mod plan;
+use plan::{
+    build_processing_plan, enforce_plan_review, ensure_output_parent_dirs, log_output_plan,
+    resolve_output_dir, ProcessingInputs, ResolvedProcessingPlan,
+};
 
 pub async fn process_payload(
     window: tauri::Window,
@@ -115,32 +95,6 @@ fn resolve_sample_rate(payload: &ProcessPayload) -> Result<audio::SampleRateConf
         .unwrap_or(audio::SampleRateConfig::Auto);
     audio::settings::validate_sample_rate_config(&sample_rate)?;
     Ok(sample_rate)
-}
-
-fn resolve_output_dir(output_dir: &str, create_if_missing: bool) -> Result<PathBuf> {
-    let base_output_dir = PathBuf::from(output_dir);
-    if create_if_missing && !base_output_dir.exists() {
-        std::fs::create_dir_all(&base_output_dir).map_err(|e| {
-            AppError::FileValidation(format!(
-                "Failed to create output directory '{}': {}",
-                sanitize_path_for_display(&base_output_dir),
-                e
-            ))
-        })?;
-    }
-    if !base_output_dir.exists() {
-        return Err(AppError::FileValidation(format!(
-            "Output directory does not exist: {}",
-            sanitize_path_for_display(&base_output_dir)
-        )));
-    }
-    if !base_output_dir.is_dir() {
-        return Err(AppError::FileValidation(format!(
-            "Output path is not a directory: {}",
-            sanitize_path_for_display(&base_output_dir)
-        )));
-    }
-    Ok(base_output_dir)
 }
 
 fn resolve_effective_processing_metadata(
@@ -216,390 +170,6 @@ fn validate_external_processing_contract_with_file_info(
     })?;
     validate_external_input_decoders(&file_info.files, &file_info.selected_decoders, &toolchain)?;
     Ok(())
-}
-
-fn resolve_collision_policy(payload: &ProcessPayload) -> CollisionPolicy {
-    payload.collision_policy.unwrap_or(CollisionPolicy::Fail)
-}
-
-fn resolve_output_kind(preview_seconds: Option<f64>) -> OutputKind {
-    if preview_seconds.is_some() {
-        OutputKind::Preview
-    } else {
-        OutputKind::Final
-    }
-}
-
-fn action_requires_output_write(action: PlannedOutputAction) -> bool {
-    matches!(
-        action,
-        PlannedOutputAction::Write
-            | PlannedOutputAction::ReplaceExisting
-            | PlannedOutputAction::RenameNew
-    )
-}
-
-fn action_is_hard_block(job: &PlannedProcessingJob) -> bool {
-    matches!(job.output.action, PlannedOutputAction::ReviewRequired)
-        && job.output.collision.as_ref().is_some_and(|collision| {
-            matches!(
-                collision.kind,
-                crate::audio::output_path::OutputCollisionKind::SourceDestinationOverlap
-                    | crate::audio::output_path::OutputCollisionKind::CanonicalPathOverlap
-            )
-        })
-}
-
-fn ensure_output_parent_dirs(plan: &ResolvedProcessingPlan) -> Result<()> {
-    for job in &plan.jobs {
-        if !action_requires_output_write(job.output.action) {
-            continue;
-        }
-        if let Some(parent) = job.output.resolved_path.parent() {
-            std::fs::create_dir_all(parent).map_err(|error| {
-                AppError::FileValidation(format!(
-                    "Cannot create output directory '{}': {}",
-                    sanitize_path_for_display(parent),
-                    error
-                ))
-            })?;
-        }
-    }
-
-    Ok(())
-}
-
-fn build_plan_signature(
-    job_type: JobType,
-    preview_seconds: Option<f64>,
-    collision_policy: CollisionPolicy,
-    jobs: &[PlannedProcessingJob],
-) -> String {
-    let mut lines = vec![
-        format!("job_type={job_type:?}"),
-        format!("preview_seconds={preview_seconds:?}"),
-        format!("collision_policy={collision_policy:?}"),
-    ];
-
-    for job in jobs {
-        let output = &job.output;
-        let collision_kind = output
-            .collision
-            .as_ref()
-            .map(|value| format!("{:?}", value.kind))
-            .unwrap_or_else(|| "none".to_string());
-        let collision_path = output
-            .collision
-            .as_ref()
-            .and_then(|value| value.conflicting_path.as_ref())
-            .map(|value| value.display().to_string())
-            .unwrap_or_default();
-        let collision_detail = output
-            .collision
-            .as_ref()
-            .and_then(|value| value.detail.clone())
-            .unwrap_or_default();
-        let output_kind = format!("{:?}", output.kind);
-        let output_action = format!("{:?}", output.action);
-        let collision_summary = format!("{collision_path}::{collision_detail}");
-        lines.push(format!(
-            "{}|{}|{}|{}|{}|{}|{}|{}",
-            job.input_index
-                .map(|value| value.to_string())
-                .unwrap_or_else(|| "merge".to_string()),
-            output.requested_path.display(),
-            output.resolved_path.display(),
-            output
-                .rename_candidate
-                .as_ref()
-                .map(|value| value.display().to_string())
-                .unwrap_or_default(),
-            output_kind,
-            output_action,
-            collision_kind,
-            collision_summary,
-        ));
-    }
-
-    lines.join("\n")
-}
-
-fn output_plan_review_message(job: &PlannedProcessingJob) -> String {
-    let destination = sanitize_path_for_display(&job.output.requested_path);
-    match job.output.collision.as_ref().map(|value| value.kind) {
-        Some(crate::audio::output_path::OutputCollisionKind::SourceDestinationOverlap)
-        | Some(crate::audio::output_path::OutputCollisionKind::CanonicalPathOverlap) => format!(
-            "Output path '{}' targets an input source file. Choose a different destination.",
-            destination
-        ),
-        _ => format!(
-            "Output collision review is required for '{}'. Re-run preflight and choose how to handle the collision.",
-            destination
-        ),
-    }
-}
-
-fn enforce_plan_review(payload: &ProcessPayload, plan: &ResolvedProcessingPlan) -> Result<()> {
-    let expected_signature = payload.preflight_signature.as_deref();
-    let current_signature = plan.plan_signature.as_str();
-
-    if let Some(signature) = expected_signature {
-        if signature != current_signature {
-            return Err(AppError::FileValidation(
-                "Output collision state changed after review. Review the collision dialog and try again."
-                    .to_string(),
-            ));
-        }
-    }
-
-    if resolve_collision_policy(payload) != CollisionPolicy::Fail && expected_signature.is_none() {
-        return Err(AppError::InvalidInput(
-            "Collision policy selections require a reviewed preflight plan.".to_string(),
-        ));
-    }
-
-    if let Some(job) = plan.jobs.iter().find(|job| action_is_hard_block(job)) {
-        return Err(AppError::FileValidation(output_plan_review_message(job)));
-    }
-
-    if let Some(job) = plan
-        .jobs
-        .iter()
-        .find(|job| job.output.action == PlannedOutputAction::ReviewRequired)
-    {
-        return Err(AppError::FileValidation(output_plan_review_message(job)));
-    }
-
-    if expected_signature.is_none() && plan.jobs.iter().any(|job| job.output.collision.is_some()) {
-        return Err(AppError::FileValidation(
-            "Output collisions require review before processing. Open the collision dialog and choose how to continue."
-                .to_string(),
-        ));
-    }
-
-    Ok(())
-}
-
-fn build_requested_output_path(
-    base_output_dir: &Path,
-    metadata: Option<&crate::metadata::AudiobookMetadata>,
-    output_naming: OutputNamingConfig,
-    source_path: Option<&Path>,
-) -> Result<PathBuf> {
-    build_output_path_preview(base_output_dir, metadata, output_naming, source_path)
-}
-
-fn build_processing_plan(
-    payload: &ProcessPayload,
-    metadata: Option<&HashMap<String, crate::metadata::MetadataIntentPatch>>,
-    inputs: &ProcessingInputs,
-) -> Result<ResolvedProcessingPlan> {
-    let job_type = payload.job_type.unwrap_or(JobType::Batch);
-    let collision_policy = resolve_collision_policy(payload);
-    let output_kind = resolve_output_kind(inputs.preview_seconds);
-    let mut claimed_outputs: HashSet<PathBuf> = HashSet::new();
-    let mut jobs = Vec::new();
-
-    match job_type {
-        JobType::Merge => jobs.push(build_merge_processing_job(
-            payload,
-            metadata,
-            inputs,
-            output_kind,
-            collision_policy,
-            &mut claimed_outputs,
-        )?),
-        JobType::Batch => jobs.extend(build_batch_processing_jobs(
-            payload,
-            metadata,
-            inputs,
-            output_kind,
-            collision_policy,
-            &mut claimed_outputs,
-        )?),
-    }
-
-    let plan_signature =
-        build_plan_signature(job_type, inputs.preview_seconds, collision_policy, &jobs);
-
-    Ok(ResolvedProcessingPlan {
-        job_type,
-        preview_seconds: inputs.preview_seconds,
-        collision_policy,
-        plan_signature,
-        jobs,
-    })
-}
-
-fn log_output_plan(phase: &str, payload: &ProcessPayload, plan: &ResolvedProcessingPlan) {
-    for job in &plan.jobs {
-        if job.output.action == PlannedOutputAction::Write
-            && job.output.collision.is_none()
-            && job.output.kind == OutputKind::Final
-        {
-            continue;
-        }
-
-        log::info!(
-            "output_plan phase={} reviewed={} policy={:?} input_index={:?} kind={:?} action={:?} requested={} resolved={} collision_kind={} collision_path={}",
-            phase,
-            payload.preflight_signature.is_some(),
-            plan.collision_policy,
-            job.input_index,
-            job.output.kind,
-            job.output.action,
-            job.output.requested_path.display(),
-            job.output.resolved_path.display(),
-            job.output
-                .collision
-                .as_ref()
-                .map(|value| format!("{:?}", value.kind))
-                .unwrap_or_else(|| "none".to_string()),
-            job.output
-                .collision
-                .as_ref()
-                .and_then(|value| value.conflicting_path.as_ref())
-                .map(|value| value.display().to_string())
-                .unwrap_or_default(),
-        );
-    }
-}
-
-fn build_merge_processing_job(
-    payload: &ProcessPayload,
-    metadata: Option<&HashMap<String, crate::metadata::MetadataIntentPatch>>,
-    inputs: &ProcessingInputs,
-    output_kind: OutputKind,
-    collision_policy: CollisionPolicy,
-    claimed_outputs: &mut HashSet<PathBuf>,
-) -> Result<PlannedProcessingJob> {
-    let paths: Vec<PathBuf> = payload.input_files.iter().map(PathBuf::from).collect();
-    let file_info = audio::get_file_list_info(&paths)?;
-    let merge_source_path = file_info
-        .files
-        .first()
-        .map(|file| file.path.clone())
-        .ok_or_else(|| {
-            AppError::InvalidInput("No input files provided for merge processing".to_string())
-        })?;
-    let merge_patch = payload
-        .input_files
-        .first()
-        .and_then(|key| metadata.and_then(|map| map.get(key)))
-        .cloned();
-    let merge_metadata =
-        resolve_effective_processing_metadata(Some(&merge_source_path), merge_patch.as_ref())?;
-    let merge_naming_metadata = resolve_naming_metadata(
-        merge_metadata.as_ref(),
-        Some(&merge_source_path),
-        merge_patch.as_ref(),
-    );
-    let requested_output = build_requested_output_path(
-        &inputs.base_output_dir,
-        merge_naming_metadata.as_ref(),
-        inputs.output_naming.clone(),
-        Some(&merge_source_path),
-    )?;
-    let source_paths = file_info
-        .files
-        .iter()
-        .map(|file| file.path.clone())
-        .collect::<Vec<_>>();
-    let output = resolve_output_plan(
-        &requested_output,
-        output_kind,
-        collision_policy,
-        claimed_outputs,
-        &source_paths,
-    )?;
-    if action_requires_output_write(output.action) {
-        claimed_outputs.insert(output.resolved_path.clone());
-    }
-
-    Ok(PlannedProcessingJob {
-        input_index: None,
-        input_path: None,
-        output,
-        metadata: merge_metadata,
-    })
-}
-
-fn build_batch_processing_jobs(
-    payload: &ProcessPayload,
-    metadata: Option<&HashMap<String, crate::metadata::MetadataIntentPatch>>,
-    inputs: &ProcessingInputs,
-    output_kind: OutputKind,
-    collision_policy: CollisionPolicy,
-    claimed_outputs: &mut HashSet<PathBuf>,
-) -> Result<Vec<PlannedProcessingJob>> {
-    if payload.input_files.is_empty() {
-        return Err(AppError::InvalidInput(
-            "No input files provided for batch processing".to_string(),
-        ));
-    }
-
-    let validated_input_paths: Vec<PathBuf> = payload
-        .input_files
-        .iter()
-        .map(|input| validate_batch_input_path(&PathBuf::from(input)))
-        .collect::<Result<_>>()?;
-
-    let mut jobs = Vec::new();
-    for (index, path) in validated_input_paths.iter().cloned().enumerate() {
-        let input = &payload.input_files[index];
-        let file_patch = metadata.and_then(|map| map.get(input)).cloned();
-        let effective_metadata =
-            resolve_effective_processing_metadata(Some(&path), file_patch.as_ref())?;
-        let naming_metadata = resolve_naming_metadata(
-            effective_metadata.as_ref(),
-            Some(&path),
-            file_patch.as_ref(),
-        );
-        let requested_output = build_requested_output_path(
-            &inputs.base_output_dir,
-            naming_metadata.as_ref(),
-            inputs.output_naming.clone(),
-            Some(&path),
-        )?;
-        let output = resolve_output_plan(
-            &requested_output,
-            output_kind,
-            collision_policy,
-            claimed_outputs,
-            &validated_input_paths,
-        )?;
-        if action_requires_output_write(output.action) {
-            claimed_outputs.insert(output.resolved_path.clone());
-        }
-        jobs.push(PlannedProcessingJob {
-            input_index: Some(index),
-            input_path: Some(path),
-            output,
-            metadata: effective_metadata,
-        });
-    }
-
-    Ok(jobs)
-}
-
-impl ResolvedProcessingPlan {
-    fn to_public(&self) -> ProcessingPreflightPlan {
-        let outputs = self
-            .jobs
-            .iter()
-            .map(|job| {
-                job.output
-                    .to_public(job.input_index, job.input_path.as_deref())
-            })
-            .collect();
-        ProcessingPreflightPlan {
-            job_type: self.job_type,
-            preview_seconds: self.preview_seconds,
-            collision_policy: self.collision_policy,
-            plan_signature: self.plan_signature.clone(),
-            outputs,
-        }
-    }
 }
 
 fn emit_batch_queue_event(
